pyproject.toml
[project]
name = "reasonify-headless"
description = "headless implementation for the Reasonify Agent"
license = { text = "MIT" }
dynamic = [ "version" ]
requires-python = ">=3.12"
authors = [ { name = "Muspi Merol", email = "me@promplate.dev" } ]
dependencies = [
  "pyodide-py~=0.26.0a4 ; sys_platform != 'emscripten'",
  "promplate~=0.3.4.5",
  "python-box~=7.1.1",
  "promptools[validation]~=0.1.3.4",
  "promplate-recipes~=0.2.2.1",
  "pyyaml~=6.0.1",
]

[build-system]
requires = [ "pdm-backend" ]
build-backend = "pdm.backend"

[tool.pdm]
build = { package-dir = "reasonify-headless" }
version = { source = "file", path = "reasonify-headless/version.py" }
distribution = true

[tool.pdm.dev-dependencies]
dev = [
  "black>=24.4.2",
  "isort>=5.13.2",
  "ruff>=0.4.4",
  "watchfiles>=0.21.0",
  "promplate-pyodide~=0.0.3.2",
  "micropip~=0.6.0",
  "webtypy~=0.1.7",
]

[tool.pdm.scripts]
fmt = { composite = [
  "ruff check --fix --exit-zero",
  "isort .",
  "black ."
] }
dev = { cmd = [
  "watchfiles",
  "pdm build --no-sdist -d static/whl",
  "reasonify-headless"
] }

[tool.black]
line-length = 150

[tool.pyright]
reportMissingModuleSource = false
pythonVersion = "3.12"
